Finished implementing Praat for articulation features

+++ articulation/articulation.py
@@ -66,7 +66,7 @@ import matplotlib.pyplot as plt
 import math
 import pysptk
 import scipy.stats as st
-from articulation_functions import extractTrans
+from articulation_functions import extractTrans, V_UV
 import uuid
 
 
@@ -78,7 +78,7 @@ import praat_functions
 
 def plot_art(data_audio,fs,F0,F1,F2,segmentsOn,segmentsOff):
     plt.figure(1)
-    plt.subplot(211)
+    plt.subplot(311)
     t=np.arange(0, float(len(data_audio))/fs, 1.0/fs)
     if len(t)>len(data_audio):
         t=t[:len(data_audio)]
@@ -89,7 +89,17 @@ def plot_art(data_audio,fs,F0,F1,F2,segmentsOn,segmentsOff):
     plt.xlim([0, t[-1]])
     plt.grid(True)
 
-    plt.subplot(212)
+    plt.subplot(312)
+    t0=np.linspace(0.0,t[-1],len(F0))
+    plt.plot(t0, F0, color='r', linewidth=2.0, label='F0')
+    plt.xlabel('Time (s)', fontsize=14)
+    plt.ylabel('Frequency (Hz)', fontsize=14)
+    plt.ylim([0,np.max(F0)+10])
+    plt.xlim([0, t0[-1]])
+    plt.grid(True)
+    plt.legend()
+
+    plt.subplot(313)
     fsp=int(len(F1)/t[-1])
     t2=np.arange(0.0, t[-1], 1.0/fsp)
     if len(t2)>len(F1):
@@ -97,6 +107,8 @@ def plot_art(data_audio,fs,F0,F1,F2,segmentsOn,segmentsOff):
     elif len(F1)>len(t2):
         F1=F1[:len(t2)]
         F2=F2[:len(t2)]
+
+
     plt.plot(t2, F1, color='k', linewidth=2.0, label='F1')
     plt.plot(t2, F2, color='g', linewidth=2.0, label='F2')
 
@@ -127,7 +139,7 @@ def plot_art(data_audio,fs,F0,F1,F2,segmentsOn,segmentsOff):
     plt.show()
 
 
-def articulation_continuous(audio_filename, flag_plots,sizeframe=0.04,step=0.02,nB=22,nMFCC=12,minf0=60,maxf0=350, voice_bias=-0.5,len_thr_miliseconds=270.0):
+def articulation_continuous(audio_filename, flag_plots,sizeframe=0.04,step=0.02,nB=22,nMFCC=12,minf0=60,maxf0=350, voice_bias=-0.5,len_thr_miliseconds=270.0, pitch_method='praat'):
 
     fs, data_audio=read(audio_filename)
     data_audio=data_audio-np.mean(data_audio)
@@ -135,22 +147,22 @@ def articulation_continuous(audio_filename, flag_plots,sizeframe=0.04,step=0.02,
     size_frameS=sizeframe*float(fs)
     size_stepS=step*float(fs)
     overlap=size_stepS/size_frameS
-    data_audiof=np.asarray(data_audio*(2**15), dtype=np.float32)
-    # F0=pysptk.sptk.rapt(data_audiof, fs, int(size_stepS), min=minf0, max=maxf0, voice_bias=voice_bias, otype='f0')
-    temp_uuid=str(uuid.uuid4().get_hex().upper()[0:6])
-    temp_filename_vuv='../tempfiles/tempVUV'+temp_uuid+'.txt'
-    temp_filename_f0='../tempfiles/tempF0'+temp_uuid+'.txt'
-    praat_functions.praat_vuv(audio_filename, temp_filename_f0, temp_filename_vuv, time_stepF0=step, minf0=minf0, maxf0=maxf0)
-    F0,_=praat_functions.decodeF0(temp_filename_f0)
-    segmentsFull,segmentsOn,segmentsOff=praat_functions.read_textgrid_trans(temp_filename_vuv,data_audio,fs,sizeframe)
-    print len(segmentsFull)
-    print len(segmentsOn)
-    print len(segmentsOff)
-    os.remove(temp_filename_vuv)
-    os.remove(temp_filename_f0)
-    # segments= read_Textgrid(path_base+'vuv.txt', file_audio, win_trans)
-    # segmentsOn=V_UV(F0, data_audio, fs, 'onset')
-    # segmentsOff=V_UV(F0, data_audio, fs, 'offset')
+
+    if pitch_method == 'praat':
+        temp_uuid=str(uuid.uuid4().get_hex().upper()[0:6])
+        temp_filename_vuv='../tempfiles/tempVUV'+temp_uuid+'.txt'
+        temp_filename_f0='../tempfiles/tempF0'+temp_uuid+'.txt'
+        praat_functions.praat_vuv(audio_filename, temp_filename_f0, temp_filename_vuv, time_stepF0=step, minf0=minf0, maxf0=maxf0)
+        F0,_=praat_functions.decodeF0(temp_filename_f0,len(data_audio)/float(fs),step)
+        segmentsFull,segmentsOn,segmentsOff=praat_functions.read_textgrid_trans(temp_filename_vuv,data_audio,fs,sizeframe)
+        os.remove(temp_filename_vuv)
+        os.remove(temp_filename_f0)
+    elif pitch_method == 'rapt':
+        data_audiof=np.asarray(data_audio*(2**15), dtype=np.float32)
+        F0=pysptk.sptk.rapt(data_audiof, fs, int(size_stepS), min=minf0, max=maxf0, voice_bias=voice_bias, otype='f0')
+        segments= read_Textgrid(path_base+'vuv.txt', file_audio, win_trans)
+        segmentsOn=V_UV(F0, data_audio, fs, 'onset')
+        segmentsOff=V_UV(F0, data_audio, fs, 'offset')
 
     BBEon, MFCCon=extractTrans(segmentsOn, fs, size_frameS, size_stepS, nB, nMFCC)
     BBEoff, MFCCoff=extractTrans(segmentsOff, fs, size_frameS, size_stepS, nB, nMFCC)
